Made project runnable from command prompt, added requirements.txt for dependency installation.
