+++ requirements.txt
@@ -1,0 +1,8 @@
+feedparser==6.0.8
+matplotlib==3.1.3
+numpy==1.18.1
+pandas==1.0.1
+PyQt5==5.15.6
+requests==2.22.0
+seaborn==0.10.0
+yfinance==0.1.67
